ensure only integer time lengths can be set

+++ break_timer/models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import User
 
 
 class Timer(models.Model):
-    work_length = models.DecimalField(decimal_places=1, max_digits=5)
-    break_length = models.DecimalField(decimal_places=1, max_digits=5)
+    work_length = models.IntegerField()
+    break_length = models.IntegerField()
     created = models.DateTimeField(auto_now_add=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='timer')
